Move all changes into extensions
